micro_web/requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cos-python-sdk-v5==1.5.5
cryptography==2.7
dicttoxml==1.7.4
Django==2.2.1
django-filter==2.0.0
django-redis==4.10.0
django-simpleui==4.0.3
djangorestframework==3.8.2
idna==2.7
lxml==4.3.4
mysqlclient==1.3.14
pycparser==2.19
pycryptodomex==3.8.1
Pygments==2.2.0
pytz==2019.1
qcloud-python-sts==3.0.5
redis==3.2.1
requests==2.19.1
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
tinify==1.5.1
urllib3==1.23
wincertstore==0.2
uwsgi==2.0.19
